# apps/backend/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Awade API",
    "description": "AI-powered educator support platform for African teachers",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/curriculum/{curriculum_id}": {
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "Get Curriculum",
        "description": "Retrieve a specific curriculum by ID.\nRequires authentication.",
        "operationId": "get_curriculum_api_curriculum__curriculum_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "curriculum_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Curriculum Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/subjects/": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "List Subjects",
        "description": "Retrieve a list of all subjects.\nRequires authentication.",
        "operationId": "list_subjects_api_subjects__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectResponse"
                  },
                  "type": "array",
                  "title": "Response List Subjects Api Subjects  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "subjects"
        ],
        "summary": "Create Subject",
        "description": "Create a new subject.\nRequires admin authentication.",
        "operationId": "create_subject_api_subjects__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/subjects/{subject_id}": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "Get Subject",
        "description": "Get a specific subject by ID.\nRequires authentication.",
        "operationId": "get_subject_api_subjects__subject_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "subjects"
        ],
        "summary": "Update Subject",
        "description": "Update a subject.\nRequires admin authentication.",
        "operationId": "update_subject_api_subjects__subject_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subject Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "subjects"
        ],
        "summary": "Delete Subject",
        "description": "Delete a subject.\nRequires admin authentication.",
        "operationId": "delete_subject_api_subjects__subject_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurriculumResponse": {
        "properties": {
          "curricula_title": {
            "type": "string",
            "title": "Curricula Title"
          },
          "country_id": {
            "type": "integer",
            "title": "Country Id"
          },
          "curricula_id": {
            "type": "integer",
            "title": "Curricula Id"
          }
        },
        "type": "object",
        "required": [
          "curricula_title",
          "country_id",
          "curricula_id"
        ],
        "title": "CurriculumResponse",
        "description": "Schema for curriculum response data."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SubjectCreate"
      },
      "SubjectResponse": {
        "properties": {
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "name"
        ],
        "title": "SubjectResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}